Cleaned up primitive evaluation scripts
Moved all functions to `primitive_metrics.jl` so evaluation can be called directly by the baseESN training scripts. `primitive_eval.jl` is now just a short call script for manually calling evaluations on datasets.
